Conflicts:
	app/controllers/Certifications.java

+++ app/manager/services/mealtickets/MealTicketStaticUtility.java
@@ -16,16 +16,17 @@ public class MealTicketStaticUtility {
 
   /**
    * Costruisce i blocchi di codici consecutivi a partire dalla lista ordinata di buoni pasto.
-   * @param mealTicketListOrdered una lista di buoni pasto ordinata
-   *     per data di scadenza e per codice blocco.
-   * @param interval intervallo da considerare.
+   *
+   * @param mealTicketListOrdered una lista di buoni pasto ordinata per data di scadenza e per
+   *                              codice blocco.
+   * @param interval              intervallo da considerare.
    * @return i blocchi
    */
   public static List<BlockMealTicket> getBlockMealTicketFromOrderedList(
-          List<MealTicket> mealTicketListOrdered, Optional<DateInterval> interval) {
+      List<MealTicket> mealTicketListOrdered, Optional<DateInterval> interval) {
 
     // TODO check della lista che sia effettivamente ordinata?
-    
+
     List<BlockMealTicket> blockList = Lists.newArrayList();
     BlockMealTicket currentBlock = null;
     MealTicket previousMealTicket = null;
@@ -36,7 +37,7 @@ public class MealTicketStaticUtility {
           && !DateUtility.isDateIntoInterval(mealTicket.date, interval.get())) {
         continue;
       }
-      
+
       //Primo buono pasto
       if (currentBlock == null) {
         previousMealTicket = mealTicket;
@@ -49,7 +50,7 @@ public class MealTicketStaticUtility {
       //Stesso blocco
       Long previous = Long.parseLong(previousMealTicket.code) + 1;
       Long actual = Long.parseLong(mealTicket.code);
-      if (previous.equals(actual) && previousMealTicket.contract.equals(mealTicket.contract) 
+      if (previous.equals(actual) && previousMealTicket.contract.equals(mealTicket.contract)
           && previousMealTicket.returned == mealTicket.returned) {
         currentBlock.getMealTickets().add(mealTicket);
       } else {
@@ -67,29 +68,30 @@ public class MealTicketStaticUtility {
     }
     return blockList;
   }
-  
+
   /**
    * La porzione di blocco associato al contratto.
+   *
    * @param blockMealTicketsOrdered i buoni pasto del blocco.
-   * @param contract contratto 
-   * @param first primo 
-   * @param last ultimo
+   * @param contract                contratto
+   * @param first                   primo
+   * @param last                    ultimo
    * @return la lista
    */
   public static List<MealTicket> blockPortion(List<MealTicket> blockMealTicketsOrdered,
       Contract contract, int first, int last) {
-    
+
     List<MealTicket> mealTickets = Lists.newArrayList();
-    Integer codeBlock = null;
+    Long codeBlock = null;
     //Controllo di consistenza.
     for (MealTicket mealTicket : blockMealTicketsOrdered) {
       if (codeBlock == null) {
         codeBlock = mealTicket.block;
       }
-      
+
       // Tutti i buoni della lista devono appartenere allo stesso blocco.
       Preconditions.checkArgument(codeBlock.equals(mealTicket.block));
-          
+
       if (mealTicket.number >= first && mealTicket.number <= last) {
         if (!mealTicket.contract.equals(contract)) {
           // un buono nell'intervallo non appartiene al contratto effettivo!!! 
@@ -101,5 +103,5 @@ public class MealTicketStaticUtility {
     }
     return mealTickets;
   }
-  
+
 }
