+++ app/models/MealTicket.java
@@ -40,7 +40,7 @@ public class MealTicket extends BaseModel {
   public LocalDate date;
 
   @Required
-  public Integer block; /*esempio 5941 3165 01 */
+  public Long block; /*esempio 5941 3165 01 */
 
   public Integer number;
